backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pyjwt
werkzeug
gunicorn
